+++ agency/cuda/detail/ndarray.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/cuda/memory/allocator.hpp>
-#include <agency/detail/ndarray.hpp>
+#include <agency/experimental/ndarray/ndarray.hpp>
 
 namespace agency
 {
@@ -13,7 +13,7 @@ namespace detail
 
 
 template<class T, class Shape = size_t, class Alloc = allocator<T>, class Index = Shape>
-using ndarray = agency::detail::ndarray<T,Shape,Alloc,Index>;
+using ndarray = agency::experimental::ndarray<T,Shape,Alloc,Index>;
 
 
 } // end detail